Move cost sensitive acquisition to core
