+++ emukit/core/acquisition/__init__.py
@@ -1,0 +1,2 @@
+from .acquisition import Acquisition
+from .acquisition_per_cost import acquisition_per_expected_cost
